requirements.txt
httpx==0.28.1
Pillow==12.1.0
playwright==1.45.0; python_version < "3.13"
playwright>=1.55.0,<2; python_version >= "3.13"
tenacity==8.2.2
typer>=0.12.3
opencv-python
aiomysql==0.2.0
redis~=4.6.0
pydantic==2.5.2; python_version < "3.13"
pydantic>=2.10.0,<3; python_version >= "3.13"
aiofiles~=23.2.1
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
jieba==0.42.1
wordcloud==1.9.3
matplotlib==3.9.0; python_version < "3.13"
matplotlib>=3.10.0,<3.11; python_version >= "3.13"
requests==2.32.3
parsel==1.9.1
pyexecjs==1.5.1
pandas==2.2.3
aiosqlite==0.21.0
pyhumps==3.8.0
cryptography>=45.0.7
alembic>=1.16.5
asyncmy>=0.2.10
sqlalchemy>=2.0.43
motor>=3.3.0
openpyxl>=3.1.2
pytest>=7.4.0
pytest-asyncio>=0.21.0